+++ src/Type/ChatBoostSourceGiveaway.php
@@ -13,6 +13,7 @@ final readonly class ChatBoostSourceGiveaway implements ChatBoostSource
         public int $giveawayMessageId,
         public ?User $user = null,
         public ?true $isUnclaimed = null,
+        public ?int $prizeStarCount = null,
     ) {}
 
     public function getSource(): string