+++ src/Type/Giveaway.php
@@ -29,5 +29,6 @@ final readonly class Giveaway
         #[ArrayMap(StringValue::class)]
         public ?array $countryCodes = null,
         public ?int $premiumSubscriptionMonthCount = null,
+        public ?int $prizeStarCount = null,
     ) {}
 }