+++ src/Type/GiveawayCreated.php
@@ -7,4 +7,10 @@ namespace Vjik\TelegramBot\Api\Type;
 /**
  * @see https://core.telegram.org/bots/api#giveawaycreated
  */
-final readonly class GiveawayCreated {}
+final readonly class GiveawayCreated
+{
+    public function __construct(
+        public ?int $prizeStarCount = null,
+    ) {
+    }
+}